Fix build to deploy, update firebase tools, set platform
Found a small error where the _TAG variable was not matching between the
build and push scripts.
In the build updated the firebase tools version to a recent version.
Put in the platform flag on the FROM line of the Dockerfile to force
to be linux/amd64. This is more of a convience so that builds will work
on Google Cloud Builds when the image is built on Apple Silicon. This
avoids getting the message
```
exec /usr/local/bin/docker-entrypoint.sh: exec format error
```